reduce dead/duplicate code
